+++ src/pages/profile/auth-layout/address-form/existed-address/schema.ts
@@ -1,0 +1,11 @@
+import * as yup from 'yup';
+
+import baseSchema from '../baseSchema';
+
+const existingAddressSchema = baseSchema.concat(
+  yup.object().shape({
+    id: yup.string().required(),
+  }),
+);
+
+export default existingAddressSchema;
